Removed debug.hpp header and renamed baseserver.h to baseserver.hpp
Updated license
