+++ license.txt
@@ -1,7 +1,7 @@
 gmsv_serversecure
 A module for Garry's Mod that mitigates exploits on the Source engine.
 -----------------------------------------------------------------------
-Copyright (c) 2015-2020, Daniel Almeida
+Copyright (c) 2015-2021, Daniel Almeida
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
